refactor screens according to ux spec
