+++ BizTechApp/App.js
@@ -2,8 +2,7 @@
 import React, {Component} from 'react';
 
 import {Platform, StyleSheet, Text, ScrollView, View} from 'react-native';
-import Register from './App/components/Register.js';
-import Profile from './App/components/Profile.js';
+// import Register from './App/components/Register.js';
 import Navigator from './App/features/Navigator';
 
 
@@ -46,4 +45,3 @@ export default App;
 		// return <EmojiDict />;
 	// }
 // }
-
